+++ DEVELOPER/CHANGES.md
@@ -8,13 +8,19 @@
 - Changed "BodilyFunctions" function prefix to "BF"
 
 - Flies now appear on human feces, whether it's on the ground or being held by the player.
-- Standing near feces now applied food sickness (nausea) and unhappiness.
+- Standing near feces now applied food sickness (nausea), unhappiness, and new "Bodily Fumes" smell moodle.
 - Human Feces can now be used as fuel for fires.
 - Human Feces can now be turned into "Throwable Human Feces", which can be thrown, picked up, thrown again, and alerts zombies. Cause why the hell not.
 
 - Added custom smell moodle (Bodily Fumes), for reacting to bodily fluids / functions (farting, poop, etc.)
 
 - Lowered time it takes to consume any of the Bodily Functions medicine types
+
+- Smell Moodle Values:
+	- Near Feces Pile: 75%
+	- Near Urine Puddle: 40%
+	- Wearing Pooped Pants: 50%
+	- Wearing Peed Pants: 30%
 
 ==============
 [VERSION 0.11.0]